Use lseek instead of lseek64 in Unix I/O

+++ src/Common/src/Interop/Unix/libc/Interop.readdir.cs
@@ -5,7 +5,7 @@ using System;
 using System.Runtime.InteropServices;
 
 using ino_t = System.IntPtr;
-using off_t = System.IntPtr;
+using off_t = System.Int64; // Assuming either 64-bit machine or _FILE_OFFSET_BITS == 64
 
 internal static partial class Interop
 {
